+++ src/navigation/VcsZoomButton.vue
@@ -38,11 +38,14 @@
     ::v-deep {
       .btn-orientation-tools {
         &:first-child {
-          box-shadow: rgba(0, 0, 0, 0.25) -2px 2px 1px -2px,
-            rgba(0, 0, 0, 0.08) 1px 0 0 0, rgba(0, 0, 0, 0.05) -1px 1px 1px 1px !important;
+          box-shadow:
+            rgba(0, 0, 0, 0.25) -2px 2px 1px -2px,
+            rgba(0, 0, 0, 0.08) 1px 0 0 0,
+            rgba(0, 0, 0, 0.05) -1px 1px 1px 1px !important;
         }
         &:last-child {
-          box-shadow: rgba(0, 0, 0, 0.25) -2px 3px 1px -2px,
+          box-shadow:
+            rgba(0, 0, 0, 0.25) -2px 3px 1px -2px,
             rgba(0, 0, 0, 0.08) 1px 4px 2px 0,
             rgba(0, 0, 0, 0.05) -1px 1px 1px 1px !important;
         }